feat(retry): add unit tests and restore retryable-request wrapper
- Restore send_retryable_request wrapper used by API-key and web-frontend
  request paths.
- Add unit tests for retry_with_backoff success, retry exhaustion, non-retryable
  short-circuit, and backoff cap.
- Cast u32 max_retries to usize at call sites.

+++ src/gemini/client.rs
@@ -314,7 +314,7 @@ impl GeminiClient {
 
         debug!(model, "sending generateContent request");
 
-        let response = crate::retry::send_retryable_request(self.max_retries, move || {
+        let response = crate::retry::send_retryable_request(self.max_retries as usize, move || {
             let client = client.clone();
             let url = url.clone();
             let request = request.clone();
@@ -348,7 +348,7 @@ impl GeminiClient {
 
         debug!(model, "sending streamGenerateContent request");
 
-        let response = crate::retry::send_retryable_request(max_retries, move || {
+        let response = crate::retry::send_retryable_request(max_retries as usize, move || {
             let client = client.clone();
             let url = url.clone();
             let request = request.clone();
